Move classifyLoads to load-counts and guard in computeLoadsByPeriodStatus
