+++ src/application/summary-logs/period-status.test.js
@@ -1078,6 +1078,7 @@ describe('computeLoadsByPeriodStatus', () => {
       wasteRecords,
       wasteBalanceRecords: wasteRecords,
       summaryLogId: SUMMARY_LOG_ID,
+      status: 'validated',
       registration: accreditedRegistration,
       processingType: 'REPROCESSOR_INPUT',
       existingRecordsMap: new Map(),
@@ -1104,6 +1105,7 @@ describe('computeLoadsByPeriodStatus', () => {
       wasteRecords: [buildWasteRecord()],
       wasteBalanceRecords: [buildWasteRecord()],
       summaryLogId: SUMMARY_LOG_ID,
+      status: 'validated',
       registration: accreditedRegistration,
       processingType: 'REPROCESSOR_INPUT',
       existingRecordsMap: new Map(),
@@ -1134,6 +1136,7 @@ describe('computeLoadsByPeriodStatus', () => {
       wasteRecords,
       wasteBalanceRecords: wasteRecords,
       summaryLogId: SUMMARY_LOG_ID,
+      status: 'validated',
       registration: accreditedRegistration,
       processingType: 'REPROCESSOR_INPUT',
       existingRecordsMap: new Map(),
@@ -1158,6 +1161,7 @@ describe('computeLoadsByPeriodStatus', () => {
       wasteRecords,
       wasteBalanceRecords: wasteRecords,
       summaryLogId: SUMMARY_LOG_ID,
+      status: 'validated',
       registration: accreditedRegistration,
       processingType: 'UNKNOWN_TYPE',
       existingRecordsMap: new Map(),
